fix: keep event display helpers import-free for content tests
Move type/message display helpers into event-display.ts so the data-URL
content-test loader can evaluate them without resolving clanka-api.
